Move macro test to macro module
